fixed: all markedDone files are shown
solution - saved array allList instead of taskList. taskList only had
tasks which were undone. When this file is saved as fileAssignments,
existing done tasks are overwritten and lost. New array allList has all
tasks - done and undone. The completed task's status was updated in this
list and saved as fileAssignment